chore: mv CertVerifierClient to separate header
