+++ filenames.gni
@@ -213,28 +213,20 @@ filenames = {
     "shell/browser/media/media_stream_devices_controller.h",
     "shell/browser/net/asar/asar_url_loader.cc",
     "shell/browser/net/asar/asar_url_loader.h",
-    "shell/browser/net/asar/url_request_asar_job.cc",
-    "shell/browser/net/asar/url_request_asar_job.h",
-    "shell/browser/net/atom_cert_verifier.cc",
-    "shell/browser/net/atom_cert_verifier.h",
-    "shell/browser/net/atom_network_delegate.cc",
-    "shell/browser/net/atom_network_delegate.h",
     "shell/browser/net/atom_url_loader_factory.cc",
     "shell/browser/net/atom_url_loader_factory.h",
     "shell/browser/net/atom_url_request_job_factory.cc",
     "shell/browser/net/atom_url_request_job_factory.h",
+    "shell/browser/net/cert_verifier_client.cc",
+    "shell/browser/net/cert_verifier_client.h",
     "shell/browser/net/proxying_url_loader_factory.cc",
     "shell/browser/net/proxying_url_loader_factory.h",
-    "shell/browser/net/js_asker.cc",
-    "shell/browser/net/js_asker.h",
     "shell/browser/net/network_context_service_factory.cc",
     "shell/browser/net/network_context_service_factory.h",
     "shell/browser/net/network_context_service.cc",
     "shell/browser/net/network_context_service.h",
     "shell/browser/net/node_stream_loader.cc",
     "shell/browser/net/node_stream_loader.h",
-    "shell/browser/net/require_ct_delegate.cc",
-    "shell/browser/net/require_ct_delegate.h",
     "shell/browser/net/resolve_proxy_helper.cc",
     "shell/browser/net/resolve_proxy_helper.h",
     "shell/browser/net/system_network_context_manager.cc",
